requirements.txt
Flask==3.0.2
boto3==1.34.46
python-dotenv==1.0.1
requests==2.28.1
python-dotenv==1.0.0
gunicorn
Werkzeug==3.1
openai==1.2.0
httpx==0.27.2
psycopg2-binary==2.9.10
websocket==0.2.1
